+++ backend/core/types/src/backend-swagger.ts
@@ -5425,6 +5425,7 @@ export enum CountyCode {
   "Alameda" = "Alameda",
   "San Mateo" = "San Mateo",
   "San Jose" = "San Jose",
+  "Detroit" = "Detroit",
 }
 
 export enum ListingEventType {